Internal Documentation

Contents

Internal Documentation

Contents

Index

Internals

Index

BlockArrays.BlockIndex

BlockArrays.blockindex2global

BlockArrays.global2blockindex

Internals

BlockArrays.BlockIndex — Type.

BlockIndex{N}

A BlockIndex is an index which stores a global index in two parts: the block and the offset index into the block.

source

BlockArrays.blockindex2global — Function.

blockindex2global{N}(block_sizes::BlockSizes{N}, block_index::BlockIndex{N}) -> inds

Converts from a block index to a tuple containing the global indices

source

BlockArrays.global2blockindex — Function.

global2blockindex{N}(block_sizes::BlockSizes{N}, inds...) -> BlockIndex{N}

Converts from global indices inds to a BlockIndex.

source