Internal Documentation

Internal Documentation

Contents

Internal Documentation

Contents

Index

Internals

Index

BlockArrays.BlockIndexRange

BlockArrays.BlockRange

BlockArrays.BlockSlice

BlockArrays.blockindex2global

BlockArrays.global2blockindex

BlockArrays.unblock

Internals

BlockArrays.blockindex2global — Function.

blockindex2global{N}(block_sizes::BlockSizes{N}, block_index::BlockIndex{N}) -> inds

Converts from a block index to a tuple containing the global indices

source

BlockArrays.global2blockindex — Function.

global2blockindex{N}(block_sizes::BlockSizes{N}, inds...) -> BlockIndex{N}

Converts from global indices inds to a BlockIndex.

source

BlockArrays.BlockRange — Type.

BlockRange(startblock, stopblock)

represents a cartesian range of blocks.

The relationship between Block and BlockRange mimicks the relationship between CartesianIndex and CartesianRange.

source

BlockArrays.BlockIndexRange — Type.

BlockIndexRange(block, startind:stopind)

represents a cartesian range inside a block.

source

BlockArrays.BlockSlice — Type.

BlockSlice(indices)

Represent an AbstractUnitRange of indices that attaches a block.

Upon calling to_indices(), Blocks are converted to BlockSlice objects to represent the indices over which the Block spans.

This mimics the relationship between Colon and Base.Slice.

source

BlockArrays.unblock — Function.

unblock(block_sizes, inds, I)

Returns the indices associated with a block as a BlockSlice.

source