Internal Documentation

Contents

Internal Documentation

Contents

Index

Internals

Index

BlockArrays.BlockIndexRange

BlockArrays.BlockRange

BlockArrays.BlockSlice

BlockArrays.BlockedUnitRange

BlockArrays.SubBlockIterator

BlockArrays.blockcolsupport

BlockArrays.blockrowsupport

BlockArrays.unblock

Internals

BlockArrays.blockcolsupport — Function

blockcolsupport(A, j)

gives an iterator containing the possible non-zero blocks in the j-th block-column of A.

BlockArrays.blockrowsupport — Function

blockrowsupport(A, k)

gives an iterator containing the possible non-zero blocks in the k-th block-row of A.

BlockArrays.BlockedUnitRange — Type

BlockedUnitRange

is an AbstractUnitRange{Int} that has been divided into blocks, and is used to represent axes of block arrays. Construction is typically via blockrange which converts a vector of block lengths to a BlockedUnitRange.

julia> blockedrange([2,2,3])
3-blocked 7-element BlockedUnitRange{Vector{Int64}}:
 1
 2
 ─
 3
 4
 ─
 5
 6
 7

BlockArrays.BlockRange — Type

BlockRange(startblock, stopblock)

represents a cartesian range of blocks.

The relationship between Block and BlockRange mimicks the relationship between CartesianIndex and CartesianRange.

BlockArrays.BlockIndexRange — Type

BlockIndexRange(block, startind:stopind)

represents a cartesian range inside a block.

BlockArrays.BlockSlice — Type

BlockSlice(indices)

Represent an AbstractUnitRange of indices that attaches a block.

Upon calling to_indices(), Blocks are converted to BlockSlice objects to represent the indices over which the Block spans.

This mimics the relationship between Colon and Base.Slice.

BlockArrays.unblock — Function

unblock(block_sizes, inds, I)

Returns the indices associated with a block as a BlockSlice.

BlockArrays.SubBlockIterator — Type

SubBlockIterator(subblock_lasts::Vector{Int}, block_lasts::Vector{Int})
SubBlockIterator(A::AbstractArray, bs::NTuple{N,AbstractUnitRange{Int}} where N, dim::Integer)

An iterator for iterating BlockIndexRange of the blocks specified by subblock_lasts. The Block index part of BlockIndexRange is determined by subblock_lasts. That is to say, the Block index first specifies one of the block represented by subblock_lasts and then the inner-block index range specifies the region within the block. Each such block corresponds to a block specified by blocklasts.

Note that the invariance subblock_lasts ⊂ block_lasts must hold and must be ensured by the caller.

Examples

julia> using BlockArrays

julia> import BlockArrays: SubBlockIterator, BlockIndexRange

julia> A = BlockArray(1:6, 1:3);

julia> subblock_lasts = axes(A, 1).lasts;

julia> @assert subblock_lasts == [1, 3, 6];

julia> block_lasts = [1, 3, 4, 6];

julia> for idx in SubBlockIterator(subblock_lasts, block_lasts)
           B = @show view(A, idx)
           @assert !(parent(B) isa BlockArray)
           idx :: BlockIndexRange
           idx.block :: Block{1}
           idx.indices :: Tuple{UnitRange}
       end
view(A, idx) = 1:1
view(A, idx) = 2:3
view(A, idx) = 4:4
view(A, idx) = 5:6

julia> [idx.block.n[1] for idx in SubBlockIterator(subblock_lasts, block_lasts)]
4-element Vector{Int64}:
 1
 2
 3
 3

julia> [idx.indices[1] for idx in SubBlockIterator(subblock_lasts, block_lasts)]
4-element Vector{UnitRange{Int64}}:
 1:1
 1:2
 1:1
 2:3